+++ src/+mic/Scan.m
@@ -107,6 +107,10 @@ classdef Scan < mic.Base
         % {logical 1x1} - true when paused
         lPaused = false;
         
+        % {logical 1x1} - true when stop command is called, set to false
+        % when start() is called
+        lStopped = true;
+        
         
         % @param {function_handle} fhSetState(stUnit, stState) - function to update the
         % N-dimensional (N-motor / N-degree-of-freedom) destination 
@@ -227,6 +231,8 @@ classdef Scan < mic.Base
             this.dSecondsElapsed = 0;
             this.ticId = tic;
             
+            this.lStopped = false;
+            
             this.go();
             
         end
@@ -274,6 +280,8 @@ classdef Scan < mic.Base
         function stop(this)
         %STOP abort the scan, reset back to start index
              this.removeClockTask();
+             this.lStopped = true;
+              
              this.u8Index = 1;
              
              %notify(this,'eAbort');
@@ -356,8 +364,11 @@ classdef Scan < mic.Base
            
             this.fhSetState(this.stUnit, this.ceValues{this.u8Index});
             
-            % Start checking the state
-            this.clock.add(@this.handleClockIsAtState, this.id(), this.dDelay); 
+            % Start checking the state.  Don't add task if we have
+            % requested a stop
+            if (~this.lStopped)
+                this.clock.add(@this.handleClockIsAtState, this.id(), this.dDelay); 
+            end
             
         end
         
@@ -378,8 +389,11 @@ classdef Scan < mic.Base
                 % Call acuire, passing in units
                 this.fhAcquire(this.stUnit, this.ceValues{this.u8Index});
                 
-                % Start checking for acquire complete
-                this.clock.add(@this.handleClockIsAcquired, this.id(), this.dDelay);
+                % Start checking for acquire complete.  Do not call if we
+                % have stopped during the acquire
+                if (~this.lStopped)
+                    this.clock.add(@this.handleClockIsAcquired, this.id(), this.dDelay);
+                end
                                 
             end