+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ﻿
 # 1.0.0-beta.25
 
+- Fixed a bug in `mic.scan` causing it to fail to stop when stop() is called during a state change or an acquisition. Also fixing a bug causing raster scanning to fail in `mic.ui.common.ScanSetup`
+
 - Generalizing scan axis and scan setup inputs
 
 - Adding getter for `mic.ui.common.Popup` to access selected value of list item.